Fix all clang-tidy findings in src/
- lexer.hpp, parser.cpp: remove std::forward on template params reused
  across loop iterations (use-after-move pattern)
- ast.hpp, expression.hpp, options.hpp, table.hpp, test_results.hpp,
  token.hpp, context.hpp: add default member initializers
  (cppcoreguidelines-pro-type-member-init /
  clang-analyzer-optin.cplusplus.UninitializedObject)
- options.cpp/hpp: convert eager static-init lambda to a function-local
  static so exceptions during construction can propagate/be caught
  instead of guaranteed std::terminate (bugprone-throwing-static-initialization)
- table.cpp/hpp, get_args.hpp, registry.hpp: fix narrowing conversions
  in iterator arithmetic via explicit std::ptrdiff_t casts
- registry.hpp, util.hpp: replace string_view::data() usage with
  explicit std::string construction where a null-terminated C-string
  is required (real correctness risk on non-null-terminated slices);
  suppress the one legitimate pointer-range usage with a documented
  NOLINTNEXTLINE
- ast.cpp: pass vector by const reference instead of by value
  (performance-unnecessary-value-param)
- log.hpp, defines.hpp: rename Log() -> log_message() and
  CWT_CONCAT_ -> CWT_CONCAT_IMPL for identifier-naming conventions
- scanner.cpp: collapse branch-clone if/else, add explicit default
  case to character-dispatch switch
- catalog.cpp, report.cpp: cast indents to int before json::dump()

Verified with full rebuild + 555 unit tests + stress-tests + example
suite all passing. Full-codebase clang-tidy scan now reports zero
findings.

+++ src/catalog.cpp
@@ -88,7 +88,7 @@ std::string as_json(std::size_t indents /* = 2 */)
     steps_catalog["types"].push_back(type_entry);
   }
 
-  return steps_catalog.dump(indents);
+  return steps_catalog.dump(static_cast<int>(indents));
 #else
   log::error(
       "nlohmann-json is not added as dependency in this build. Can not "
